Version1.17 更新部分：1.完成了赠礼部分的内容，玩家可以通过赠送NPC喜爱的物品内容去提高双方的关系和好感度，从而取得更多的正面效果。2.初步开发了玩家的技能使用效果和UI界面的功能。3.新增了技能的熟练度部分的逻辑，包括数值随熟练度等级变化以及熟练度的增长。后续可能进一步随着技能系统的开发和完善进行更改。 修复部分：1.修改了玩家的互动逻辑，让战斗和和平模式分离开同时应用于UI上的变化，让玩家可以更明确的识别当前状态。2.修改了玩家的攻击列表和目标逻辑，引入unity自带的physic2D资源包所封装的代码进行检测，让敌人检索逻辑更加清晰和直接。3. 修改了NPC和敌方单位之间的攻击逻辑，让技能和普通攻击的AI攻击逻辑更加完善。 预告部分：1.开发技能部分的逻辑和效果（包括玩家和NPC）以及应用在动画和生成物上。2.使用Unity Shader制作技能范围提示和选择范围提示让技能系统更加完善。3.在技能系统的基础上开发buff系统使得二者相辅相成并结合战斗系统将整个战斗模块的逻辑完善。

+++ Assets/Scripts/Utilities/DataCollection.cs
@@ -162,7 +162,7 @@ public class RelationShip
     //FIXME: 后期视情况改成是否拥有自己的社交圈
     //public string Name;
     [Range(-100, 100)]
-    public int Favourability;
+    public float Favourability;
 }
 
 [System.Serializable]
